fix(testing): update FakeAnsiblePort signatures to align with trait
Updates `FakeAnsiblePort::tags_by_role` and `role_for_tag` signatures to match
recent refactoring in the `AnsiblePort` trait that returns references instead of
owned values, fixing the CI build failures.
