+++ src/app/commands/config/mod.rs
@@ -74,7 +74,7 @@ fn deploy_internal(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::testing::{ansible::FakeAnsiblePort, fs::FakeFsPort};
+    use crate::testing::fakes::{FakeAnsiblePort, FakeFsPort};
     use std::path::PathBuf;
 
     #[test]
