+++ src/app/commands/deploy_configs.rs
@@ -85,7 +85,7 @@ pub fn copy_dir_recursive(src: &Path, dst: &Path, fs: &dyn FsPort) -> Result<(),
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::testing::{ansible::FakeAnsiblePort, fs::FakeFsPort};
+    use crate::testing::fakes::{FakeAnsiblePort, FakeFsPort};
     use std::path::PathBuf;
 
     #[test]
