+++ src/testing/mod.rs
@@ -1,4 +1,3 @@
 //! In-process test doubles and builders.
 
-pub mod ansible;
-pub mod fs;
+pub mod fakes;
